Add base game content for most games
