+++ content/games/csgo.mdx
@@ -4,8 +4,27 @@ icon: "../img/games/csgo/icon.svg"
 banner: "../img/games/csgo/banner.png"
 logo: "../img/games/csgo/logo.svg"
 slug: csgo
+contacts:
+  - name: Chris Newman
+    discord: Chris#7693
+links:
+  - text: Rules coming soon
+    icon: book
+    href: "/rules/csgo.pdf"
+    class: "btn-secondary"
+    disabled: true
+  - text: Smash.gg Page
+    icon: smashgg
+    href: "https://smash.gg/tournament/gamefest-2019/events/counter-strike-global-offensive-5v5-byoc"
+    class: "btn-secondary"
+  - text: Register
+    icon: ticket-alt
+    href: "/register/"
+    class: "btn-primary"
 ---
 
 #### Tournament Overview
 
 Details will be available soon
+
+<Contact contacts={props.contacts} />
