+++ content/games/dota.mdx
@@ -4,6 +4,20 @@ icon: "../img/games/dota/icon.svg"
 banner: "../img/games/dota/banner.png"
 logo: "../img/games/dota/logo.svg"
 slug: dota
+links:
+  - text: Rules coming soon
+    icon: book
+    href: "/rules/dota.pdf"
+    class: "btn-secondary"
+    disabled: true
+  - text: Smash.gg Page
+    icon: smashgg
+    href: "https://smash.gg/tournament/gamefest-2019/events/dota-2-5v5-byoc"
+    class: "btn-secondary"
+  - text: Register
+    icon: ticket-alt
+    href: "/register/"
+    class: "btn-primary"
 ---
 
 #### Tournament Overview
