+++ content/games/fighting-games.mdx
@@ -4,8 +4,28 @@ icon: "../img/games/fighting-games/icon.svg"
 banner: "../img/games/fighting-games/banner.jpg"
 logo: "../img/games/fighting-games/logo.svg"
 slug: fighting-games
+contacts:
+  - name: Cassandra "doojeeboy" Naomi
+    discord: cat#4477
+links:
+  - text: Register
+    icon: ticket-alt
+    href: "/register/"
+    class: "btn-primary"
 ---
 
 #### Tournament Overview
 
-Details will be available soon
+Gamefest 2019 will have a variety of different fighting game tournaments, with Melee/Ultimate tournaments occurring on **the second day, November 3<sup>rd</sup>** and all other fighting game tournaments occurring on the **first day, November 2<sup>nd</sup>**
+
+##### Games list
+
+- Smash Ultimate
+- Smash Melee
+- Tekken
+- BBtag
+- Guilty Gear
+- UNIST
+- DBFZ
+
+<Contact contacts={props.contacts} />
