+++ content/games/hearthstone.mdx
@@ -4,6 +4,11 @@ icon: "../img/games/hearthstone/icon.png"
 banner: "../img/games/hearthstone/banner.jpg"
 logo: "../img/games/hearthstone/logo.png"
 slug: hearthstone
+contacts:
+  - name: Bhanu "unahb" Garg
+    discord: unahb#8581
+  - name: Marc "Twinov" Hoeltge
+    discord: Twinov#6441
 links:
   - text: Rules coming soon
     icon: book
@@ -12,7 +17,7 @@ links:
     disabled: true
   - text: Smash.gg Page
     icon: smashgg
-    href: "https://smash.gg/tournament/gamefest-2019/events/hearthstone-conquest/overview"
+    href: "https://smash.gg/tournament/gamefest-2019/events/hearthstone-conquest"
     class: "btn-secondary"
   - text: Register
     icon: ticket-alt
@@ -35,3 +40,5 @@ The tournament will use the Conquest format in which each round is a best of fiv
 
 - 4 standard format decks required
 - 1 ban
+
+<Contact contacts={props.contacts} />
