+++ content/games/league.mdx
@@ -4,8 +4,41 @@ icon: "../img/games/league/icon.png"
 banner: "../img/games/league/banner.jpg"
 logo: "../img/games/league/logo.png"
 slug: league
+contacts:
+  - name: Eugene "doojeeboy" Ahn
+    discord: Doojeeboy#5805
+links:
+  - text: Rules coming soon
+    icon: book
+    href: "/rules/league.pdf"
+    class: "btn-secondary"
+    disabled: true
+  - text: Smash.gg Page
+    icon: smashgg
+    href: "https://smash.gg/tournament/gamefest-2019/events/league-of-legends-5v5-byoc"
+    class: "btn-secondary"
+  - text: Register
+    icon: ticket-alt
+    href: "/register/"
+    class: "btn-primary"
 ---
 
 #### Tournament Overview
 
-Details will be available soon
+Gamefest 2019 will have an League of Legends tournament that will span **both days** and will feature the following:
+
+- 5v5, team-based
+- Bring-your-own-computer
+- Tournament played on Days 1 and 2 (November 2<sup>nd</sup> and 3<sup>rd</sup>, respectively), schedule permitting
+  - Finals will occur on the second day
+
+##### Competitive Structure
+
+The tournament format will consist of Swiss-style on Saturday followed by Single Elimination playoffs on Sunday:
+
+- **Map:** Summoner’s Rift
+- **Team Size:** 5
+- **Allow Spectators:** Lobby Only
+- **Game Type:** Tournament Draft
+
+<Contact contacts={props.contacts} />
