+++ content/games/rocket-league.mdx
@@ -4,6 +4,20 @@ icon: "../img/games/rocket-league/icon.svg"
 banner: "../img/games/rocket-league/banner.jpg"
 logo: "../img/games/rocket-league/logo.svg"
 slug: rocket-league
+links:
+  - text: Rules coming soon
+    icon: book
+    href: "/rules/rocket-league.pdf"
+    class: "btn-secondary"
+    disabled: true
+  - text: Smash.gg Page
+    icon: smashgg
+    href: "https://smash.gg/tournament/gamefest-2019/events/rocket-league-3v3-byoc"
+    class: "btn-secondary"
+  - text: Register
+    icon: ticket-alt
+    href: "/register/"
+    class: "btn-primary"
 ---
 
 #### Tournament Overview
