+++ src/scss/main.scss
@@ -207,10 +207,6 @@ article {
     }
 }
 
-section p:last-of-type {
-    margin-bottom: 0;
-}
-
 .top-banner {
     padding-top: 4rem;
     padding-bottom: 2rem;
